# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BetterPlay API",
        "description": "Aquí podrás ver todas las funciones de la API detalladamente",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/friends/friendRequest": {
            "put": {
                "tags": [
                    "friends"
                ],
                "summary": "Crea una petición de amistad",
                "description": "Recibe la id de otro user y le manda una petición de amistad",
                "operationId": "21990fbabe7cec6efb6ef364f2f3376b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizado Correctamente"
                    }
                }
            }
        },
...
        "/api/notifications/create": {
            "put": {
                "tags": [
                    "notifications"
                ],
                "summary": "Crea una notificación",
                "description": "Recibe un nombre de la notificción y el tipo y la crea",
                "operationId": "88999a952fc8e8755ebce2329e0ac0a9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "text": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizado Correctamente"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}